The code has been broken into chunks for easier maintenance. New code has been written to allow performing crosshand phase calibration using A-team source. An image based method has been used. But since we will only use flux of the source I have extracted the DS using the visibilities themselves.
