+++ solar_realtime_pipeline.py
@@ -41,6 +41,9 @@ from ovrolwasolar import visualization as ovis
 from ovrolwasolar import refraction_correction as orefr
 from ovrolwasolar import coords as ocoords
 
+import data_downloader
+
+
 #import gc # garbage collection
 #gc.enable()
 
@@ -70,53 +73,7 @@ def set_memory_limit(percentage=0.1):
     soft, hard = resource.getrlimit(resource.RLIMIT_AS)
     resource.setrlimit(resource.RLIMIT_AS, (int(get_memory() * 1024 * percentage), hard))
 
-def source_riseset(skycoord, date_time,observatory='ovro', altitude_limit=15):
-    '''
-    :param date_time: input time in astropy.time.Time format
-    :param observatory: name of the observatory recognized by astropy
-    :param altitude_limit: lower limit of altitude to consider. Default to 15 degrees.
-    :param skycoord: Source can be provided by using a astropy skycoordinate. 
-                     
-    :return trise, tset  ## has 30 min resolution
-    '''
-    try:
-        date_mjd = Time(date_time).mjd
-    except Exception as e:
-        logging.error(e)
 
-    obs = EarthLocation.of_site(observatory)
-    t0 = Time(int(date_mjd), format='mjd')
-    source_rise=t0
-    
-
-    alt = skycoord.transform_to(AltAz(obstime=t0, location=obs)).alt.degree
-
-    source_risen=False
-    if alt>altitude_limit:
-        source_risen=True
-
-    if not source_risen:    
-        while alt < altitude_limit:
-            source_rise += TimeDelta(1800., format='sec')
-            alt = skycoord.transform_to(AltAz(obstime=source_rise, location=obs)).alt.degree
-            
-        source_set=source_rise+TimeDelta(1800., format='sec')
-        while alt > altitude_limit:
-            source_set += TimeDelta(1800., format='sec')
-            alt = skycoord.transform_to(AltAz(obstime=source_set, location=obs)).alt.degree
-        return source_rise,source_set
-    else:
-        while alt > altitude_limit:
-            source_rise -= TimeDelta(1800., format='sec')
-            alt = skycoord.transform_to(AltAz(obstime=source_rise, location=obs)).alt.degree
-        
-        source_set=t0
-        alt = skycoord.transform_to(AltAz(obstime=source_set, location=obs)).alt.degree
-        while alt>altitude_limit:
-            source_set += TimeDelta(1800., format='sec')
-            alt = skycoord.transform_to(AltAz(obstime=source_set, location=obs)).alt.degree
-
-        return source_rise,source_set
     
 
 def sun_riseset(date=Time.now(), observatory='ovro', altitude_limit=15.):
@@ -154,448 +111,10 @@ def sun_riseset(date=Time.now(), observatory='ovro', altitude_limit=15.):
     return t0, t1
 
 
-def list_msfiles(intime, lustre=True, file_path='slow', server=None, time_interval='10s', 
-                 bands=['32MHz', '36MHz', '41MHz', '46MHz', '50MHz', '55MHz', '59MHz', '64MHz', '69MHz', '73MHz', '78MHz', '82MHz']):
-    """
-    Return a list of visibilities to be copied for pipeline processing for a given time
-    :param intime: astropy Time object
-    :param server: name of the server to list available data
-    :param lustre: if True, specific to lustre system on lwacalim nodes. If not, try your luck in combination with file_path
-    :param file_path: file path to the data files. For lustre, it is either 'slow' or 'fast'. For other servers, provide full path to data.
-    :param time_interval: Options are '10s', '1min', '10min'
-    :param bands: bands to list/download. Default to 12 bands above 30 MHz. Full list of available bands is
-            ['13MHz', '18MHz', '23MHz', '27MHz', '32MHz', '36MHz', '41MHz', '46MHz', '50MHz', '55MHz', '59MHz', '64MHz', '69MHz', '73MHz', '78MHz', '82MHz']
-    """
-    intimestr = intime.isot[:-4].replace('-','').replace(':','').replace('T','_')
-    datestr = intime.isot[:10]
-    hourstr = intime.isot[11:13]
-    if time_interval == '10s':
-        tstr = intimestr[:-1]
-    if time_interval == '1min':
-        tstr = intimestr[:-2]
-    if time_interval == '10min':
-        tstr = intimestr[:-3]
-
-    msfiles = []
-    if lustre:
-        processes=[]
-        for b in bands:
-            pathstr = '/lustre/pipeline/{0:s}/{1:s}/{2:s}/{3:s}/'.format(file_path, b, datestr, hourstr)
-            if server:
-                cmd = 'ssh ' + server + ' ls ' + pathstr + ' | grep ' + tstr
-            else:
-                cmd = 'ls ' + pathstr + ' | grep ' + tstr
-            p = subprocess.Popen(cmd, stdout=subprocess.PIPE, shell=True)
-            processes.append(p)
-            filenames = p.communicate()[0].decode('utf-8').split('\n')[:-1]
-            if len(filenames) > 0:
-                for filename in filenames:
-                    if filename[-6:] == 'MHz.ms':
-                        filestr = pathstr + filename
-                        tmpstr = filename[:15].replace('_', 'T')
-                        timestr = tmpstr[:4] + '-' + tmpstr[4:6] + '-' + tmpstr[6:11] + ':' + tmpstr[11:13] + ':' + tmpstr[13:]
-                        freqstr = filename[16:21]
-                        msfiles.append({'path': filestr, 'name': filename, 'time': timestr, 'freq': freqstr})
-            else:
-                logging.info('Did not find any files at the given time {0:s}.'.format(intime.isot)) 
-    else:
-        if server:
-            cmd = 'ssh ' + server + ' ls ' + file_path + ' | grep ' + tstr
-        else:
-            cmd = 'ls ' + file_path + ' | grep ' + tstr
-        p = subprocess.run(cmd, capture_output=True, shell=True)
-        filenames = p.stdout.decode('utf-8').split('\n')[:-1]
-        if len(filenames) > 0:
-            for filename in filenames:
-                if filename[-6:] == 'MHz.ms':
-                    if server:
-                        pathstr = '{0:s}:{1:s}/{2:s}'.format(server, file_path, filename)
-                    else:
-                        pathstr = '{0:s}/{1:s}'.format(file_path, filename)
-                    tmpstr = filename[:15].replace('_', 'T')
-                    timestr = tmpstr[:4] + '-' + tmpstr[4:6] + '-' + tmpstr[6:11] + ':' + tmpstr[11:13] + ':' + tmpstr[13:]
-                    freqstr = filename[16:21]
-                    msfiles.append({'path': pathstr, 'name': filename, 'time': timestr, 'freq': freqstr})
-        else:
-            logging.info('Did not find any files at the given time {0:s}.'.format(intime.isot)) 
-            msfiles = []
-    return msfiles
-
-def download_msfiles_cmd(msfile_path, server, destination):
-    if server:
-        p = subprocess.Popen(shlex.split('rsync -az --numeric-ids --info=progress2 --no-perms --no-owner --no-group {0:s}:{1:s} {2:s}'.format(server, msfile_path, destination)))
-    else:
-        #p = subprocess.Popen(shlex.split('rsync -az --numeric-ids --info=progress2 --no-perms --no-owner --no-group {0:s} {1:s}'.format(msfile_path, destination)))
-        p = subprocess.Popen(shlex.split('cp -r {0:s} {1:s}'.format(msfile_path, destination)))
-    std_out, std_err = p.communicate()
-    if std_err:
-        print('<<',Time.now().isot,'>>',std_err)
-
-def download_msfiles(msfiles, destination='/fast/solarpipe/realtime_pipeline/slow_working/', bands=None, verbose=True, server=None, maxthread=3):
-    from multiprocessing.pool import ThreadPool
-    """
-    Parallelized downloading for msfiles returned from list_msfiles() to a destination.
-    """
-    inmsfiles_path = [f['path'] for f in msfiles]
-    inmsfiles_name = [f['name'] for f in msfiles]
-    inmsfiles_band = [f['freq'] for f in msfiles]
-    omsfiles_path = []
-    omsfiles_name = []
-    if bands is None:
-        omsfiles_path = inmsfiles_path
-        omsfiles_name = inmsfiles_name
-    else:
-        for bd in bands:
-            if bd in inmsfiles_band:
-                idx = inmsfiles_band.index(bd)
-                #omsfiles_server.append(inmsfiles_server[idx])
-                omsfiles_path.append(inmsfiles_path[idx])
-                omsfiles_name.append(inmsfiles_name[idx])
-
-    nfile = len(omsfiles_path)
-    if nfile == 0:
-        print('<<',Time.now().isot,'>>','No files to download. Abort...')
-        return -1
-    time_bg = timeit.default_timer() 
-    if verbose:
-        print('<<',Time.now().isot,'>>','I am going to download {0:d} files'.format(nfile))
-
-    tp = ThreadPool(maxthread)
-    for omsfile_path in omsfiles_path:
-        tp.apply_async(download_msfiles_cmd, args=(omsfile_path, server, destination))
-
-    tp.close()
-    tp.join()
-
-    time_completed = timeit.default_timer() 
-    if verbose:
-        print('<<',Time.now().isot,'>>','Downloading {0:d} files took in {1:.1f} s'.format(nfile, time_completed-time_bg))
-    omsfiles = [destination + n for n in omsfiles_name]
-    return omsfiles
 
 
-def download_timerange(starttime, endtime, download_interval='1min', destination='/fast/solarpipe/20231027/slow/', 
-                server=None, lustre=True, file_path='slow', bands=None, verbose=True, maxthread=5):
-    '''
-    :param download_interval: If str should either be 10s, 1min or 10min. If integer should be in seconds.
-    '''
-    time_bg = timeit.default_timer() 
-    t_start = Time(starttime)
-    t_end = Time(endtime)
-    print('<<',Time.now().isot,'>>','Start time: ', t_start.isot)
-    print('<<',Time.now().isot,'>>','End time: ', t_end.isot)
-    if not os.path.exists(destination):
-        os.makedirs(destination)
-
-    if download_interval == '10s':
-        dt = TimeDelta(10., format='sec')
-    elif download_interval == '1min':
-        dt = TimeDelta(60., format='sec')
-    elif download_interval == '10min':
-        dt = TimeDelta(600., format='sec')
-    else:
-        try:
-            download_interval=int(download_interval)
-            if download_interval%10!=0:
-                logging.warning("Data is recorded with 10s cadence. Separation should be a multiple of 10."+\
-                                "Setting download interval to nearest multiple of 10")
-            dt=TimeDelta(download_interval,format='sec')
-        except Exception as e:
-            logging.error("download interval should either be 10s, 1min, 10min, or an integer in seconds")
-            print ("download interval should either be 10s, 1min, 10min, or an integer in seconds")
-            raise e
-    nt = int(np.ceil((t_end - t_start) / dt))
-    if isinstance(download_interval, int):
-        download_interval=str(download_interval)+"s"
-    print('====Will download {0:d} times at an interval of {1:s}===='.format(nt, download_interval))
-    for i in range(nt):
-        intime = t_start + i * dt
-        msfiles = list_msfiles(intime, server=server, lustre=lustre, file_path=file_path, time_interval='10s', bands=bands)
-        if verbose:
-            print('Downloading time ', intime.isot)
-        download_msfiles(msfiles, destination=destination, bands=bands, verbose=verbose, server=server, maxthread=maxthread)
-    time_completed = timeit.default_timer() 
-    if verbose:
-        print('====Downloading all {0:d} times took {1:.1f} s===='.format(nt, time_completed-time_bg))
 
 
-def download_calibms(calib_time, download_fold = '/lustre/solarpipe/realtime_pipeline/ms_calib/', doflag=True,
-        bands=['13MHz', '18MHz', '23MHz', '27MHz', '32MHz', '36MHz', '41MHz', '46MHz', '50MHz', '55MHz', '59MHz', '64MHz', '69MHz', '73MHz', '78MHz', '82MHz']):
-    """
-    Function to download calibration ms files for all or selected bands based on a given time
-    :param calib_time: time selected for generating the calibration tables. A string recognized by astropy Time format
-    :param download_fold: directory to hold the downloaded ms files 
-    :param bands: band selection. Default to use all 16 bands.
-    """
-    if type(calib_time) == str:
-        try:
-            calib_time = Time(calib_time)
-        except:
-            print('The input time needs to be astropy.time.Time format')
-    print(socket.gethostname(), '=======Calibrating Time {0:s}======='.format(calib_time.isot))
-    ms_calib0 = list_msfiles(calib_time, file_path='slow', bands=bands)
-    ms_calib = download_msfiles(ms_calib0, destination=download_fold, bands=bands)
-    ms_calib.sort()
-    if doflag:
-        for ms_calib_ in ms_calib:
-            try:
-               antflagfile = flagging.flag_bad_ants(ms_calib_)
-            except:
-                exc_type, exc_obj, exc_tb = sys.exc_info()
-                fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-                logging.error(exc_type, fname, exc_tb.tb_lineno)
-    return ms_calib
-
-
-def gen_caltables(calib_in, bcaltb=None, uvrange='>10lambda', refant='202', flag_outrigger=True, 
-        proc_dir='./',doplot=True):
-    """
-    Function to generate calibration tables for a list of calibration ms files
-    :param calib_in: input used for calibration. This can be either a) a string of time stamp recognized by astropy.time.Time 
-            or b) a specific list of ms files used for calibration
-    :param bcaltb: name of the calibration tables. Use the default if None
-    :param uvrange: uv range to be used, default to '>10lambda'
-    :param refant: reference antenna, default to '202'
-    :param flag_outrigger: if True, flag all outrigger antennas. These would be used for beamforming.
-    :proc_dir: directory to process the data and hold output files.
-    """
-    import pandas as pd
-    download_fold = proc_dir + '/ms_calib/'
-    caltable_fold = proc_dir + '/caltables/'
-    beam_caltable_fold = proc_dir + '/caltables_beam/'
-
-    if not os.path.exists(download_fold):
-        os.makedirs(download_fold)
-
-    if not os.path.exists(caltable_fold):
-        os.makedirs(caltable_fold)
-
-    if not os.path.exists(beam_caltable_fold):
-        os.makedirs(beam_caltable_fold)
-
-    bcaltbs = []
-    chan_freqs = []
-    if type(calib_in) == list:
-        calib_in.sort()
-        ms_calib = calib_in
-    elif type(calib_in) == str:
-        try:
-            calib_time = Time(calib_in)
-        except:
-            print('<<',Time.now().isot,'>>','The input time needs to be astropy.time.Time format. Abort...')
-            return -1
-        ms_calib = download_calibms(calib_time, doflag=True, download_fold=download_fold)
-    else:
-        print('<<',Time.now().isot,'>>','Input not recognized. Abort...')
-        return -1
-
-    
-    if len(ms_calib) > 0:
-        # TODO: somehow the parallel processing failed if flagging has run. I have no idea why. Returning to the slow serial processing.
-        #pool = multiprocessing.pool.Pool(processes=len(ms_calib))
-        #gen_calib_partial = partial(calibration.gen_calibration, uvrange=uvrange, caltable_fold=caltable_fold,
-        #                refant=refant)
-        #result = pool.map_async(gen_calib_partial, ms_calib)
-        #timeout = 2000.
-        #result.wait(timeout=timeout)
-        #bcaltbs = result.get()
-        #pool.close()
-        #pool.join()
-        for ms_calib_ in ms_calib:
-            try:
-                bcaltb = calibration.gen_calibration(ms_calib_, uvrange=uvrange, caltable_fold=caltable_fold, refant=refant)
-                msmd.open(ms_calib_)
-                chan_freqs.append(msmd.chanfreqs(0))
-                msmd.done()
-                bcaltbs.append(bcaltb)
-            except Exception as e:
-                print('<<',Time.now().isot,'>>','Something is wrong when making calibrations for ', ms_calib_)
-                print(e)
-            
-        chan_freqs = np.concatenate(chan_freqs)
-        if doplot:
-            create_waterfall_plot(bcaltbs,ms_calib)
-    else:
-        print('<<',Time.now().isot,'>>','The list of calibration ms files seems to be empty. Abort...')
-        return -1
-        
-
-    if flag_outrigger:
-        core_ant_ids, exp_ant_ids = flagging.get_antids(ms_calib[0])
-        bcaltbs_bm = []
-        bmcalfac = []
-        for bcaltb in bcaltbs:
-            bcaltb_bm = beam_caltable_fold + '/' + os.path.basename(bcaltb)
-            os.system('cp -r ' + bcaltb + ' ' + bcaltb_bm)
-            tb.open(bcaltb_bm, nomodify=False)
-            flags = tb.getcol('FLAG')
-            npol, nch, nant = flags.shape
-            for exp_ant_id in exp_ant_ids:
-                flags[:, :, exp_ant_id] = True
-            num_ant_per_chan = nant - np.sum(flags, axis=2)
-            bmcalfac_per_chan = num_ant_per_chan ** 2.
-            bmcalfac.append(bmcalfac_per_chan)
-            tb.putcol('FLAG', flags)
-            tb.close()
-            bcaltbs_bm.append(bcaltb_bm)
-
-        bmcalfac = np.concatenate(bmcalfac, axis=1)
-        # write channel frequencies and corresponding beam scaling factors into a csv file
-        df = pd.DataFrame({"chan_freqs":chan_freqs, "calfac_x":bmcalfac[0], "calfac_y":bmcalfac[1]})
-        bcalfac_file = beam_caltable_fold + '/' + os.path.basename(bcaltb)[:15] + '_bmcalfac.csv'
-        df.to_csv(bcalfac_file, index=False)
-        return bcaltbs, bcaltbs_bm, bcalfac_file
-    else:
-        return bcaltbs
-
-def get_gain_amplitude_phase(caltable):
-    tb=table()
-    tb.open(caltable)
-    try:
-        data=tb.getcol('CPARAM')
-        flag=tb.getcol('FLAG')
-    finally:
-        tb.close()
-
-    pos=np.where(flag==True)
-    data[pos]=np.nan
-    return np.abs(data),np.angle(data)
-
-def get_caltable_freq(caltable):
-    tb=table()
-    tb.open(os.path.join(caltable,"SPECTRAL_WINDOW"))
-    try:
-        chan_freq=tb.getcol('CHAN_FREQ').flatten()
-    finally:
-        tb.close()
-    return chan_freq
-
-def robust_linear_fit(x,y,num_trial=5,thresh=3):
-    for i in range(num_trial):
-        pos=np.where(np.isnan(y)==False)
-        poly=np.polyfit(x[pos],y[pos],deg=1)
-        predicted_y=np.poly1d(poly)(x)
-        residual=predicted_y-y
-        std=np.nanstd(residual)
-        pos=np.where(residual>thresh*std)
-        y[pos]=np.nan
-    return poly
-    
-def find_delay(freqs,phase):
-    pos=np.where(phase<0)[0]
-    if pos.size>0:
-        phase+=2*np.pi ### assume that phase is from -pi to pi
-
-    pos=np.where((np.isnan(phase)==False) & (freqs>40*1e6))[0]
-    if pos.size==0:
-        return np.nan
-    phase1=phase[pos]
-    freqs1=freqs[pos]/1e6 ### converting to MHz
-    unwrapped_phase=np.unwrap(phase1)
-    delay=robust_linear_fit(freqs1,unwrapped_phase)[0]/(2*np.pi) ### delay is in microseconds
-    return delay
-    
-def find_delay_all_ant_corr(freqs,phase):
-    phase_shape=phase.shape
-    num_corr=phase_shape[0]
-    num_ant=phase_shape[2]
-    
-    delay=np.zeros((num_corr,num_ant))*np.nan
-    
-    for i in range(num_corr):
-        for j in range(num_ant):
-            wrapped_phase=phase[i,:,j]
-            delay[i,j]=find_delay(freqs,wrapped_phase)
-    return delay
-    
-def create_waterfall_plot(caltables,msnames,figname=None,num_chan=192,num_ant=352):
-    
-    bands=['13MHz', '18MHz', '23MHz', '27MHz', '32MHz', '36MHz', '41MHz', \
-            '46MHz', '50MHz', '55MHz', '59MHz', '64MHz', '69MHz', '73MHz', \
-            '78MHz', '82MHz']
-            
-    msnames.sort()
-    
-    assert len(caltables)<=len(msnames),"The number of MS is smaller than the number of caltables. Please check"
-
-    num_msnames=len(msnames)
-    num_bands=len(bands)
-    
-    ms_freqs_str=[]
-    caltable_freqs=np.zeros(len(caltables))
-    
-    amp=np.zeros((2,num_bands*num_chan,num_ant))
-    phase=np.zeros((2,num_bands*num_chan,num_ant))
-    freqs=np.zeros((num_bands*num_chan))
-
-    for j,msname in enumerate(msnames):
-        ms_freqs_str.append(utils.get_freqstr_from_name(msname))
-    
-    for j,caltable in enumerate(caltables):
-        caltable_freqs[j]=get_caltable_freq(caltable)[0]/1e6 ### converting to MHz
-    
-
-    j=0   
-    for k,band in enumerate(bands):
-        if band!=ms_freqs_str[j]:
-            print ("MS file for "+band+" is missing")
-            amp[:,k*num_chan:(k+1)*num_chan,:]=np.nan
-            phase[:,k*num_chan:(k+1)*num_chan,:]=np.nan
-            freqs[k*num_chan:(k+1)*num_chan]=np.nan
-            continue    
-        msname=msnames[j]
-        ms_freq_MHz=int(ms_freqs_str[j].split('MHz')[0])
-        ind=np.argmin(abs(ms_freq_MHz-caltable_freqs))
-        #ind=caltable_freqs.index(ms_freqs[j])
-        if abs(ms_freq_MHz-caltable_freqs[ind])>2:
-            print ("Freq index not found. Some caltables have not been produced")
-            amp[:,k*num_chan:(k+1)*num_chan,:]=np.nan
-            phase[:,k*num_chan:(k+1)*num_chan,:]=np.nan
-            freqs[k*num_chan:(k+1)*num_chan]=np.nan
-        else:    
-            amp[:,k*num_chan:(k+1)*num_chan,:], phase[:,k*num_chan:(k+1)*num_chan,:]=\
-                                                            get_gain_amplitude_phase(caltables[ind])
-            freqs[k*num_chan:(k+1)*num_chan]=get_caltable_freq(caltables[ind])
-        j+=1
-    
-    delays=find_delay_all_ant_corr(freqs,phase)
-        
-    fig,ax=plt.subplots(nrows=2,ncols=1,sharex=True,figsize=[15,10])
-
-    for ind,pol in zip([0,1],['XX','YY']):
-        im=ax[ind].imshow(amp[ind,:,:],origin='lower',interpolation='none',vmax=0.005,\
-                            vmin=0.0002,cmap='viridis',extent=(0,351,13,87),aspect='auto')
-        ax[ind].set_title(pol)
-        plt.colorbar(im,ax=ax[ind])
-
-    fig.suptitle(utils.get_timestr_from_name(caltables[0]))
-    fig.supxlabel("Antenna")
-    fig.supylabel("Gain amplitude")
-
-    if not isinstance(figname,list):
-        figname_amp='_'.join(caltables[0].split('_')[:-1])+"_amp.png"
-    else:
-        figname_amp=figname[0]
-    plt.savefig(figname_amp)
-    plt.close()
-    
-    fig,ax=plt.subplots(nrows=2,ncols=1,sharex=True,figsize=[12,10])
-    
-    for ind,pol in zip([0,1],['XX','YY']):
-        im=ax[ind].plot(delays[ind],'ro')
-        ax[ind].set_title(pol)
-
-    fig.suptitle(utils.get_timestr_from_name(caltables[0]))
-    fig.supxlabel("Antenna")
-    fig.supylabel("Delay (microseconds)")
-    if not isinstance(figname,list):
-        figname_delay='_'.join(caltables[0].split('_')[:-1])+"_delay.png"
-    else:
-        figname_delay=figname[1]
-    plt.savefig(figname_delay)
-    plt.close()
-    
-    return
         
 def convert_caltables_for_fast_vis(solar_ms,calib_ms,caltables):
     fast_caltables=[]
@@ -1084,16 +603,16 @@ def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=
         print('<<',Time.now().isot,'>>')
         print(socket.gethostname(), '=======Processing Time {0:s}======='.format(image_time.isot))
         #logging.info('=======Processing Time {0:s}======='.format(image_time.isot))
-        msfiles0 = list_msfiles(image_time, lustre=lustre, server=server, file_path=file_path, time_interval='10s')
+        msfiles0 = data_downloader.list_msfiles(image_time, lustre=lustre, server=server, file_path=file_path, time_interval='10s')
         
         if len(msfiles0) < len(bands):
             # try to find missing times from nearby times that are +-4 s within the reference time
-            msfiles0_before = list_msfiles(image_time - TimeDelta(10., format='sec'), lustre=lustre, server=server, file_path=file_path, time_interval='10s')
+            msfiles0_before = data_downloader.list_msfiles(image_time - TimeDelta(10., format='sec'), lustre=lustre, server=server, file_path=file_path, time_interval='10s')
             if len(msfiles0_before) > 0:
                 msfiles0_before_ts = [f['time'] for f in msfiles0_before]
             else:
                 msfiles0_before_ts = []
-            msfiles0_after = list_msfiles(image_time + TimeDelta(10., format='sec'), lustre=lustre, server=server, file_path=file_path, time_interval='10s')
+            msfiles0_after = data_downloader.list_msfiles(image_time + TimeDelta(10., format='sec'), lustre=lustre, server=server, file_path=file_path, time_interval='10s')
             if len(msfiles0_after) > 0:
                 msfiles0_after_ts = [f['time'] for f in msfiles0_after]
             else:
@@ -1179,7 +698,7 @@ def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=
             print('<<',Time.now().isot,'>>','==Copying file over to working directory==')
             logging.debug('====Copying file over to working directory====')
             time1 = timeit.default_timer()
-            msfiles = download_msfiles(msfiles0, destination=visdir_work, bands=bands)
+            msfiles = data_downloader.download_msfiles(msfiles0, destination=visdir_work, bands=bands)
             time2 = timeit.default_timer()
             logging.debug('Time taken to copy files is {0:.1f} s'.format(time2-time1))
 
@@ -1360,8 +879,10 @@ def pipeline_quick(image_time=Time.now() - TimeDelta(20., format='sec'), server=
                     if not os.path.exists(fig_mfs_dir_sub_synop):
                         os.makedirs(fig_mfs_dir_sub_synop)
                     
-                     
-                    fits_images, plotted_image = compress_plot_images(fitsfiles, btime, datedir, imagedir_allch_combined, hdf_dir, \
+                    fitsfiles.sort()
+                    allstokes_fits=combine_pol_images(fitsfiles,stokes) 
+                    utils.correct_primary_beam(allstokes_fits,pol=stokes)
+                    fits_images, plotted_image = compress_plot_images(allstokes_fits, btime, datedir, imagedir_allch_combined, hdf_dir, \
                                             fig_mfs_dir, stokes, fast_vis=fast_vis)
                     
                     logging.info("Level 1 images plotted ok")
@@ -1495,8 +1016,39 @@ def image_times(msfiles_slfcaled, imagedir_allch, nch_out=12, stokes='I', beam_f
                 
     return fitsfiles
 
+def combine_pol_images(fitsfiles,stokes):
+    pols=stokes.split(',')
+    multi_pol_fitsfiles=[]
+    stokes_order=''
+    
+    stokes_images={}
+    for pol in pols:
+        stokes_images[pol]=[]
+    
+    for img in fitsfiles:
+        pol=img.split('-')[-2]
+        stokes_images[pol].append(img)
+        
+    if len(pols)!=1:
+        for j,pol in enumerate(pols):
+            files=stokes_images[pol]
+            files.sort()
+            multi_pol_fitsfiles.append(files)
+            stokes_order+=pol
+        stokes_order=','.join(stokes_order)
+        num_files=len(multi_pol_fitsfiles[0])
+        multi_pol_fitsfiles=np.array(multi_pol_fitsfiles)
+        fitsfiles_pol_combined=[None]*num_files
+        for j in range(num_files):
+            outfits=utils.combine_IQUV_images(multi_pol_fitsfiles[:,j].tolist(),stokes_order=stokes_order)
+            fitsfiles_pol_combined[j]=outfits
+    return fitsfiles_pol_combined
+
 def compress_plot_images(fitsfiles, starttime, datedir, imagedir_allch_combined, \
                             hdf_dir, fig_mfs_dir, stokes, fast_vis=False):    
+                            
+    
+        
 
     if fast_vis:
         imagename_pre = 'ovro-lwa-48'
